check if we're updating

+++ setup.sh
@@ -4,6 +4,7 @@
 cd "$( dirname "${BASH_SOURCE[0]}" )"
 
 #symlinks
+update=${1:-no}
 PWD=`pwd`
 PREFIX='.'
 DOTFILES=$(git ls-files)
@@ -12,17 +13,17 @@ IGNOREFILES=( setup.sh README.md .gitignore )
 # Cleanup old installations
 
 # remove antigen
-if [ -e "${HOME}/.antigen" ]; then
+if [ -e "${HOME}/.antigen" && "$update" == "no"]; then
     rm -rf ${HOME}/.antigen
 fi
 
 # remove zgen
-if [ -e "${HOME}/.zgen" ]; then
+if [ -e "${HOME}/.zgen" && "$update" == "no" ]; then
     rm -rf ${HOME}/.zgen
 fi
 
 # remove vim-plug
-if [ -e "${HOME}/.vim/autoload/plug.vim" ]; then
+if [ -e "${HOME}/.vim/autoload/plug.vim" && "$update" == "no" ]; then
     rm -f ${HOME}/.vim/autoload/plug.vim
 fi
 
@@ -46,8 +47,8 @@ for DOTFILE in ${DOTFILES}; do
     [[ -d ${SYMLINKDIR} ]] || mkdir -p ${SYMLINKDIR}
     ln -sf ${PWD}/${DOTFILE} ${SYMLINK}
 
-    if [ -e "${PWD}/${DOTFILE}.local" ]; then
-        echo "Adding local settings from ${PWD}/${DOTFILE}.local => ${SYMLINK}"
-        cat ${PWD}/${DOTFILE}.local >> ${SYMLINK}
+    if [ -e "${PWD}/${dotfile}.local" && "$update" == no ]; then
+        echo "Adding local settings from ${PWD}/${dotfile}.local => ${symlink}"
+        cat ${PWD}/${dotfile}.local >> ${symlink}
     fi
 done
